break reportedcomment and changestatusmodal into own components
